suaDieta/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.3
defusedxml==0.7.1
Django==4.1.1
django-allauth==0.51.0
django-bootstrap-form==3.4
django-crispy-forms==1.14.0
django-multiselectfield==0.1.12
django-tempus-dominus==5.1.2.17
django-widget-tweaks==1.4.12
idna==3.4
oauthlib==3.2.2
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.6.0
python3-openid==3.2.0
requests==2.28.1
requests-oauthlib==1.3.1
sqlparse==0.4.2
tzdata==2022.2
urllib3==1.26.12
